fix infra repository class name

+++ nts.uk/pr/proto/nts.uk.pr.proto.infra/src/main/java/nts/uk/ctx/pr/proto/infra/entity/paymentdata/QpdmtPaydayPK.java
@@ -25,7 +25,7 @@ public class QpdmtPaydayPK implements Serializable {
 
 	@Basic(optional = false)
 	@Column(name = "PROCESSING_YM")
-	public BigDecimal processingYM;
+	public int processingYM;
 
 	@Basic(optional = false)
 	@Column(name = "SPARE_PAY_ATR")
