+++ nts.uk/pr/proto/nts.uk.pr.proto.infra/src/main/java/nts/uk/ctx/pr/proto/infra/entity/paymentdata/personalcommute/PprmtPersonCommutePK.java
@@ -21,6 +21,6 @@ public class PprmtPersonCommutePK implements Serializable {
 	public String pId;
 	
 	@Column(name = "STR_YM")
-	public BigDecimal strYm;
+	public int strYm;
 
 }
